+++ Tp1_labo/src/operaciones.h
@@ -8,4 +8,4 @@ int suma(int *numA, int *numB, int *resSuma);
 int resta(int *numA, int *numB, int *resResta);
 int producto(int *numA, int *numB, int *resProd);
 int division(int *numA, int *numB, float *resDiv);
-int factoreo(int num, long *resFact);
+int factoreo(int num, int *resFact);
